Fix print formats

+++ src/serial_io.c
@@ -302,7 +302,8 @@ void prepare_array_serial(
     h_err = H5Pset_chunk(h_prop, rank, chunk_shape);
     if (h_err < 0)
       error("Error while setting chunk size (%llu, %llu) for field '%s'.",
-            chunk_shape[0], chunk_shape[1], props.name);
+            (unsigned long long)chunk_shape[0],
+            (unsigned long long)chunk_shape[1], props.name);
   }
 
   /* Are we imposing some form of lossy compression filter? */
